move flashpoll specific templatetags to flashpoll app
