+++ euth/dashboard/templates/euth_dashboard/includes/project_form_result_flashpoll_tab.html
@@ -1,4 +1,4 @@
-{% load i18n form_tags widget_tweaks %}
+{% load i18n form_tags flashpoll_form_tags widget_tweaks %}
 
 <script src="http://code.highcharts.com/js/highcharts.js"></script>
 <script type="text/javascript">
@@ -53,14 +53,14 @@ $(function () {
     <div class="form-section">
         <h3 class="form-section-heading">{% trans 'Result '%}</h3>
         <div class="form-section-fields">
-		
+
             <div class="form-group{% if field.errors %} has-error{% endif %}">
                 <label for="{{form.project.result.id_for_label}}">{{ form.project.result.label }}</label>
                 <p class='form-help-text'>{{ form.project.result.help_text }}</p>
                 {% render_field form.project.result class="form-control" %}
                 <span class="help-block">{{ error }}</span>
-            </div>			
-			{% if pollresult.noOfReceivedResults != 0 %}			
+            </div>
+			{% if pollresult.noOfReceivedResults != 0 %}
 				{% for question in poll.questions %}
 					{% if question.questionType == 'CHECKBOX' or question.questionType == 'RADIO' %}
 					<div class="form-group">
@@ -73,7 +73,7 @@ $(function () {
 					</div>
 					{% endif %}
 				{% endfor %}
-			{% endif %}			
+			{% endif %}
         </div>
     </div>
