+++ euth/dashboard/templates/euth_dashboard/project_form.html
@@ -1,5 +1,5 @@
 {% extends "euth_dashboard/base_organisation.html" %}
-{% load static widget_tweaks i18n form_tags%}
+{% load static widget_tweaks i18n form_tags flashpoll_form_tags %}
 {% block dashboard_content %}
 <h1 class="black-divider dashboard-title">{{ heading }} {% if view.blueprint %}{{ view.blueprint.title }}{% endif %}</h1>
 
@@ -25,7 +25,6 @@
         <button type="submit" name="save_draft" class="btn btn-sm btn-default">{% trans 'unpublish' %}</button>
         {% endif %}
         <button type="submit" name="publish" class="btn btn-sm btn-primary">{% trans 'save and publish' %}</button>
-        
         {% if module_settings == "euth_flashpoll" %}
             {% with poll=form.module_settings.poll.value|get_poll_from_string %}
             {% if project.pk != None and poll.preview == False %}
@@ -35,7 +34,7 @@
         {% else %}
             {% if project.pk != None %}
             <button type="button" class="btn btn-sm btn-danger"  data-toggle="modal" data-target="#deleteModal">delete</button>
-            {% endif %}        
+            {% endif %}
         {% endif %}
 
     </form>
